separate parsing lexer output&IR
