+++ route_verification/bgp/src/cmp/peering.rs
@@ -1,5 +1,3 @@
-use parse::*;
-
 use super::*;
 
 pub struct CheckPeering<'a> {
