+++ route_verification/lex/src/lib.rs
@@ -3,28 +3,23 @@ use std::{collections::BTreeMap, mem};
 use log::{debug, error};
 use serde::{Deserialize, Serialize};
 
-pub mod action;
 pub mod ast;
-pub mod community;
 pub mod filter;
 pub mod lines;
 pub mod mp_import;
 pub mod peering;
 pub mod rpsl_object;
-pub mod stats;
 #[cfg(any(test, feature = "test_util"))]
 pub mod test_util;
 #[cfg(test)]
 pub mod tests;
 
 pub use {
-    action::Actions,
     ast::Ast,
-    community::Call,
     filter::Filter,
     lines::{expressions, io_wrapper_lines, lines_continued, rpsl_objects, RPSLObject, RpslExpr},
-    mp_import::Versions,
+    mp_import::{Casts, Entry, PeeringAction, Versions},
     peering::{AsExpr, ComplexAsExpr, Peering},
     rpsl_object::{AsOrRouteSet, AutNum, FilterSet, PeeringSet},
-    stats::Counts,
+    shared_struct::{action::*, community::Call, stats::Counts},
 };
