+++ route_verification/parse/src/action.rs
@@ -1,14 +1,9 @@
-use ::lex::action;
-
 use super::*;
 
 /// <https://www.rfc-editor.org/rfc/rfc2622#section-6.1.1>
-pub type Actions = BTreeMap<String, Action>;
-
-pub fn parse_actions(actions: action::Actions) -> Actions {
-    actions
-}
-
 /// <https://www.rfc-editor.org/rfc/rfc2622#section-7>
 /// For now, we do not further parse `<action>`s.
-pub type Action = action::Action;
+#[inline]
+pub fn parse_actions(actions: Actions) -> Actions {
+    actions
+}
