+++ route_verification/shared_struct/src/community.rs
