+++ route_verification/shared_struct/src/lib.rs
@@ -1,0 +1,5 @@
+use serde::{Deserialize, Serialize};
+
+pub mod action;
+pub mod community;
+pub mod stats;
